requirements.txt
aio-pika==6.8.0
aiohttp==3.5.4
aioredis==2.0.0
aiormq==3.3.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.6
chardet==3.0.4
click==8.0.1
colorama==0.4.4
dataclasses==0.8
django-ckeditor==6.1.0
django-js-asset==1.2.2
fastapi==0.68.0
h11==0.12.0
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.6.3
multidict==4.7.6
pamqp==2.3.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==1.0.2
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
yarl==1.6.3
zipp==3.5.0
